# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "https://github.com/junun/admin-go",
        "contact": {
            "name": "Junun",
            "url": "https://github.com/junun/admin-go",
            "email": "junun717@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/admin/config/app": {
            "get": {
                "description": "应用列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用配置"
                ],
                "summary": "应用列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "active 查询项目是否生效",
                        "name": "active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "环境类型id",
                        "name": "envId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "应用新增",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用配置"
                ],
                "summary": "应用新增",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "应用信息",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.AppResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/config/app/{id}": {
            "put": {
                "description": "应用修改",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用配置"
                ],
                "summary": "应用修改",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "应用ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "应用信息",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.AppResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "应用删除",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用配置"
                ],
                "summary": "应用删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "应用ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/config/deploy": {
            "get": {
                "description": "应用模板列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用配置"
                ],
                "summary": "应用模板列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "通过app id查询",
                        "name": "aid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "应用发布模板新增",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用配置"
                ],
                "summary": "应用发布模板新增",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "应用发布模板信息",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.HostRoleResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.AppResource": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "integer"
                },
                "deployType": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string"
                },
                "enableSync": {
                    "type": "integer"
                },
                "envId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tid": {
                    "type": "integer"
                }
            }
        },
...
        "admin.HostRoleResource": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}